delete wasm
