+++ wasm/light_client.go
